Move tenant releated files to tenant directory
