+++ pageserver/src/lib.rs
@@ -13,9 +13,7 @@ pub mod profiling;
 pub mod repository;
 pub mod task_mgr;
 pub mod tenant;
-pub mod tenant_config;
-pub mod tenant_mgr;
-pub mod tenant_tasks;
+
 pub mod trace;
 pub mod virtual_file;
 pub mod walingest;
@@ -26,6 +24,8 @@ pub mod walredo;
 use std::path::Path;
 
 use tracing::info;
+
+use crate::tenant::tenant_config;
 
 use crate::task_mgr::TaskKind;
 
@@ -56,7 +56,7 @@ pub async fn shutdown_pageserver(exit_code: i32) {
 
     // Shut down all the tenants. This flushes everything to disk and kills
     // the checkpoint and GC tasks.
-    tenant_mgr::shutdown_all_tenants().await;
+    tenant::tenant_mgr::shutdown_all_tenants().await;
 
     // Stop syncing with remote storage.
     //
